feat: add cache to bitmap download
dfgdfg
