+++ app/src/main/java/ani/dantotsu/widgets/upcoming/UpcomingRemoteViewsFactory.kt
@@ -12,6 +12,7 @@ import ani.dantotsu.connections.anilist.Anilist
 import ani.dantotsu.media.Media
 import ani.dantotsu.settings.saving.PrefManager
 import ani.dantotsu.settings.saving.PrefName
+import ani.dantotsu.util.BitmapUtil.Companion.downloadImageAsBitmap
 import ani.dantotsu.util.BitmapUtil.Companion.roundCorners
 import ani.dantotsu.util.Logger
 import com.google.gson.GsonBuilder
@@ -182,33 +183,6 @@ class UpcomingRemoteViewsFactory(private val context: Context) :
 
         return rv
     }
-
-    private fun downloadImageAsBitmap(imageUrl: String): Bitmap? {
-        var bitmap: Bitmap? = null
-        var inputStream: InputStream? = null
-        var urlConnection: HttpURLConnection? = null
-
-        try {
-            val url = URL(imageUrl)
-            urlConnection = url.openConnection() as HttpURLConnection
-            urlConnection.requestMethod = "GET"
-            urlConnection.connect()
-
-            if (urlConnection.responseCode == HttpURLConnection.HTTP_OK) {
-                inputStream = urlConnection.inputStream
-                bitmap = BitmapFactory.decodeStream(inputStream)
-            }
-        } catch (e: Exception) {
-            e.printStackTrace()
-        } finally {
-            inputStream?.close()
-            urlConnection?.disconnect()
-        }
-        return bitmap?.let { roundCorners(it) }
-    }
-
-
-
 
     override fun getLoadingView(): RemoteViews {
         return RemoteViews(context.packageName, R.layout.item_upcoming_widget)
